Fix build break on Solaris
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@63337 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ lwadvapi/unthreaded/memory/lwstr.c
@@ -921,7 +921,7 @@ LwURLDecodeString(
     {
         if (pIn[i] == '%')
         {
-            if (!isxdigit(pIn[i+1]) || !isxdigit(pIn[i+2]))
+            if (!isxdigit((int)(pIn[i+1])) || !isxdigit((int)(pIn[i+2])))
             {
                 dwError = ERROR_INVALID_PARAMETER;
                 BAIL_ON_LW_ERROR(dwError);
